move lint handling to an extends class in test_timeline; add separate unit test & better error output for lint errors
